Created query with Retrofit
Updated .gitignore for ProtectedClass that holds API key.

Api.java is an interface - this has a void method, activeListings() that
will run the @GET call through Retrofit, passing in the method call
/listings/active.
Inside the activeListings() method, we annotate a @QUERY that will deal
with additional parameters.
Finally, we set a Callback so that the ActiveListings object will be
automatically filled.

In Etsy.java, we have the getApi() method that will return a RestAdapter
from Retrofit. Here, we set the endpoint for the Etsy site.
We also set a RequestInterceptor that will intercept the request,
appending our API_KEY for Etsy.

Finally, we'll build and create the RestAdapter, using the Api class as
the interface.