Created ListingAdapter
After we get our data from the Etsy API using Retrofit, we'll then be
brought to this method. Here, we'll set the activeListings through the
success() method and notifyDataSetChanged().

Nested a ListingHolder class that will be the ViewHolder. This will just
store all of the views together. This just saves us having too many
findViewById() calls that could prove costly in terms of performance.

In ListingAdapter's constructor, we'll get the context and our
LayoutInflater from the context.
In onCreateViewHolder(), we'll create a ListingHolder and inflate our
views using the inflater.
In onBindViewHolder(), we'll get the current listing and set everything.
In getItemCount(), check we don't have null and return the length.