UI Preparations
Added our RecyclerView, ProgressBar and TextView for an error. Put all
of these in a FrameLayout (Draws child views in a stack).

In MainActivity, added references to the views and created 3 new
methods:
showLoading() - Shows the progressBar only.
showList() - Shows the list only.
showError() - shows the error only.

showLoading() is called when the app first runs.