Displaying the data in the RecyclerView!
In ListingAdapter.java, fixed the fact that the inflater was public.
Updated our constructor to take in MainActivity as the Context instead.
Will also show the list in the success() call via MainActivity.
In failure(), will show the error via MainActivity.
Added new getActiveListings() method that will return activeListings.

In MainActivity.java, added a key for STATE_ACTIVE_LISTINGS. Here, we'll
set up the RecyclerView using a StaggeredGridLayoutManager with 1 column
that scrolls vertically.
We'll then set the adapter to a new ListingAdapter, passing in the
context.

Our data will be stored in the savedInstanceState -
onSaveInstanceState() will be ran as part of the super method at the
start of onCreate().
Here, we will get the active listings using our new method and check for
null. If they're not null, we can go ahead and put the parcelable data
in using the key STATE_ACTIVE_LISTINGS. If it is null, we'll not do
anything here and deal with that in onCreate().

After SaveInstanceState(), we'll check for null. If it is null, show the
loading indicator and get our listings from the API.
If we have our listings in the SavedInstanceState, run the success()
method on the adapter that will pass in the data and show the list and
notifyDataSetChanged().

Finally, in Etsy.java, we also have a new getActiveListings() method
that will call the API to get the listings. This is different to the
getActiveListings() method in ListingAdapter.java that will return the
current values of the listings (Which may or may not be null).