Created new GoogleServicesHelper
This class will implement the required methods for the GoogleApiClient.

Created GoogleServicesListener interface that we'll use for the main UI.

Next, we have our constructor for the class that takes in our Activity
and GoogleServicesListener.

We'll then build a GoogleApiClient in this constructor, with the
ConnectionCallbacks() and OnConnectionFailedListener() being part of
this class.

onConnected(), onConnectionSuspended() and onConnectionFailed() are all
implemented methods that will be used later.