Handling states in GoogleApiClient
Added request codes at top of class.

Added two new methods, connect() and disconnect() that will run the
appropriate methods off of the apiClient.

Added isGooglePlayServicesAvailable() method that will check if we can
use Google Play Services or if there's a problem.
If there is a problem, we'll forward that over to the
GooglePlayServicesUtil that will try to resolve the issue. We'll still
return false just in case the problem was not fixed, however.