Checking if Google Play Services is available
Using isGooglePlayServicesAvailable() to see if we should
connect/disconnect to the apiClient or tell the listener that Google
Play Services is not available (Disable the listener).

In onConnectionFailed(), we have access to the ConnectionResult and we
can use this to see if there's a resolution to the users problem.
If there is a resolution, try to run it. Even if there was a problem
starting the resolution, try connecting still as this method might be
called multiple times and maybe we can connect through a different path
this time.
If there is no resolution however, let's just disconnect.

In handleActivityResult(), we will handle the result from our
update/disabled/invalid issues with REQUEST_CODE_AVAILABILITY or
problems with the resolution with REQUEST_CODE_RESOLUTION.
Here, we can check to see if the resultCode that gets passed in is
RESULT_OK, which would allow us to connect again.
If there's still a problem, we'll just have to disconnect.