Integrating GoogleServicesHelper in MainActivity
ListingAdapter also implements
GoogleServicesHelper.GoogleServicesListener.

Set up new isGooglePlayServicesAvailable boolean that will be used at
top scope. Defaulted to false.

Created if/else statement that will deal with inflating the Google+
button should Google Play Services be available. We'll deal with this
later.

Created new onConnected() and onDisconnected() methods where we'll check
if our item count is 0 (We need to get the listings) and then
notifyDataSetChanged() after setting the boolean for availability.

In MainActivity, created new GoogleServicesHelper object which is set in
onCreate() after creating the adapter.

Updated savedInstanceState() check. We'll now show the loading at the
start before the check and we'll check for not null. If it isn't null,
we'll then update the listings. We don't need to worry about null here
as it is dealt with during the GoogleServicesHelper object creation.
Added new onStart() and onStop() methods that connect/disconnect the API
connection.

Added new onActivityResult() method where we'll handle the result in
GoogleServicesHelper.