Got a working Google+ button!
Added ic_social_share icons.

In layout_listing.xml, set up the PlusOneButton and ImageButton that
uses the ic_social_share icon.

In ListingAdapter, when Google Play Services are available, we'll show
the plusOneButton, initialize it with our URL and tell it to not display
any annotations.
If Google Play Services is not available, we'll remove the
plusOneButton.

Added the plusOneButton and shareButton to our ListingHolder.

In MainActivity, will call notifyDataSetChanged() in onActivityResult()
so that we can show the redrawn version with the Google+ buttons, if
available.