Added Share Button
In MainActivity.java, added check to only update the display if our
requestCode is from the +1 button.

In ListingAdapter.java, now using this new REQUEST_CODE for all +1
calls.

Also set up the share button - if Google Play Services are available,
we'll use the PlusShare Builder. If not, we'll use the generic Share
Intent.
Both will be started for a result.