Final UI Improvements
In ListingAdapter.java, set an OnClickListener() for the ImageView. This
will open up the URL for the listing via an Intent.

New highlight_selector XML's - these will be used as the highlight that
appears when the user holds down on a View.
On Lollipop, we use the ripple.

In layout_listing.xml, wrapped everything in a FrameLayout and set the
foreground as our highlight_selector.
Set minimum height for ImageView and set scaleType to centerCrop.
Added 4dp padding for all text.
Updated text colors.

Added new colors in colors.xml

Updated styles.xml for new UI design.